Enhance task management by implementing authorization checks in TaskController and CompleteTaskController using Gate; update TaskPolicy to allow user-specific actions; modify PHPUnit configuration to enable debugging during tests.

+++ laravel_api/app/Http/Controllers/Api/V2/TaskController.php
@@ -3,10 +3,17 @@
 namespace App\Http\Controllers\Api\V2;
 
 use App\Models\Task;
+use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
+use App\Http\Resources\TaskResource;
+use Illuminate\Support\Facades\Gate;
 use App\Http\Requests\StoreTaskRequest;
 use App\Http\Requests\UpdateTaskRequest;
-use App\Http\Resources\TaskResource;
+
+// Note that the index and store methods do not require a model instance.
+// In these case, you need to pass a class name instead of model instance.
+
+
 
 class TaskController extends Controller
 {
@@ -15,8 +22,15 @@ class TaskController extends Controller
      */
     public function index()
     {
+        Gate::authorize('viewAny', Task::class);
+
+        // if ($request->user()->cannot('viewAny', Task::class)) {
+        //     abort(403);
+        // }  Remember note ▲
         // return TaskResource::collection(Task::all()); //
-        return request()->user()->tasks()->toResourceCollection();
+        return TaskResource::collection(request()->user()->tasks);
+        // return TaskResource::collection(request()->user()->tasks()->get());
+
     }
 
     /**
@@ -24,6 +38,11 @@ class TaskController extends Controller
      */
     public function store(StoreTaskRequest $request)
     {
+        Gate::authorize('create', Task::class);
+
+        // if ($request->user()->cannot('create', Task::class)) {
+        //     abort(403);
+        // }
         $task = $request->user()->tasks()->create($request->validated());
         return $task->toResource();
     }
@@ -31,8 +50,13 @@ class TaskController extends Controller
     /**
      * Display the specified resource.
      */
-    public function show(Task $task)
+    public function show(Task $task, Request $request)
     {
+        Gate::authorize('view', $task);
+
+        // if ($request->user()->cannot('view', $task)) {
+        //     abort(403);
+        // }
         // return new TaskResource($task);//
         // return TaskResource::make($task);//
         return $task->toResource();
@@ -43,6 +67,11 @@ class TaskController extends Controller
      */
     public function update(UpdateTaskRequest $request, Task $task)
     {
+        Gate::authorize('update', $task);
+
+        // if ($request->user()->cannot('update', $task)) {
+        //     abort(403);
+        // }
         $task->update($request->validated());
         return $task->toResource();
     }
@@ -50,8 +79,12 @@ class TaskController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Task $task)
+    public function destroy(Task $task, Request $request)
     {
+        Gate::authorize('delete', $task);
+        // if ($request->user()->cannot('delete', $task)) {
+        //     abort(403);
+        // }
         $task->delete();
         return response()->noContent();
     }
